Refactoring serializers and schema into dedicated api directory in ntags, inventory and botany
