+++ backend/botany/api/views.py
@@ -2,7 +2,7 @@ from rest_framework import viewsets
 from rest_framework_simplejwt.authentication import JWTAuthentication
 from rest_framework.permissions import IsAuthenticatedOrReadOnly
 
-from .models import UserPlant
+from ..models import UserPlant
 from .serializers import UserPlantSerializer
