+++ backend/botany/viewsets.py
@@ -1,25 +1,7 @@
 from wagtail.admin.panels import TabbedInterface, FieldPanel, FieldRowPanel, InlinePanel, ObjectList
-from wagtail.snippets.views.chooser import SnippetChooserViewSet
 from wagtail.snippets.views.snippets import SnippetViewSet
 
 from .models import UserPlant
-
-
-class UserPlantChooserViewSet(SnippetChooserViewSet):
-    model = UserPlant
-    icon = "plant"
-    choose_one_text = "Choose a plant"
-    edit_item_text = "Edit this plant"
-    form_fields = ["name", "description"]
-
-    def __init__(self, *args, **kwargs):
-        # import requests
-        # from django.conf import settings
-        # r = requests.get(f"{settings.WAGTAILADMIN_BASE_URL}/api/users/")
-        # r.raise_for_status()
-        # results = r.json()
-        # return results
-        super().__init__(*args, **kwargs)
 
 
 class UserPlantViewSet(SnippetViewSet):
@@ -35,7 +17,6 @@ class UserPlantViewSet(SnippetViewSet):
     admin_url_namespace = "plants"
     base_url_path = "inventory/plants"
     add_to_admin_menu = True
-    chooser_viewset_class = UserPlantChooserViewSet
 
     detail_panels = [
         FieldRowPanel([
