+++ backend/botany/wagtail_hooks.py
@@ -1,7 +1,13 @@
 from wagtail import hooks
 from wagtail.snippets.models import register_snippet
 
+from .views import plant_species_chooser_viewset
 from .viewsets import UserPlantViewSet
+
+
+@hooks.register("register_admin_viewset")
+def register_plant_species_chooser_viewset():
+    return plant_species_chooser_viewset
 
 
 @hooks.register("register_icons")
