+++ backend/ntags/api/serializers.py
@@ -1,6 +1,6 @@
 from rest_framework import serializers
 
-from .models import NFCTag, NFCTagScan
+from ..models import NFCTag, NFCTagScan
 
 
 class NFCTagScanSerializer(serializers.ModelSerializer):
